Base branch HIFIO-CS-ES commits- squashed
HadoopInputFormatIO with junits

Removed HadoopInputFormatIOConstants file and removed the use of constants from related classes.

Added comment for InputFormat ParameterizedType

Added Cassandra and Elastic tests/ITs on in HadoopInputFormatIO code.

Removed Writable Coder from HadoopInputFormat and referred it from Hadoop Common module

Restructured hadoop-input-format project for jdk1.8

Removed unwanted classes

Removed separate package of unit tests input, Moved them to inputformat package as per comments from iemejia

Removed .toDelete file and test.txt

Changes in resources files, added apache license , comments , renamed files as per Stephens comments, added comment to pom for child module

Created hifio child module for basic HadoopInputFormatIO code

Added dataflow runner profile in io pom

Updated teardown.sh

Removed the unused src folder

Updated version to 0.7.0

Removed the extra files created during rebase

Updated elasticsearch kubernetes scripts

Updated the getProgress() method to avoid progress value less than 0 or 1

Added comments in Kubernetes scripts

Cleaned up hifio pom.xml

Added while loop for sleep

Changes in poms- clean up, removal of version properties of sl4j, added df runner profile in hif test pom etc.

Reverted the change of Dataflow runner profile

Added seperate script for ycsb tool

Updated names of the script files

Added separate script for installing python

Removed <requireMavenVersion> tag from enforcer

Updated Javadoc

Removed the success message.

group id changed for parent as org.apache.beam

Refactoring -rename directory hadoop-input-format -> hadoop
 input-format,jdk1.8-tests

formatting correction for new module hadoop, renamed from hadoop-input-format to hadoop

formatting correction for new module hadoop, renamed from hadoop-input-format to hadoop, Removed annotation VisibleForTesting from IO class

Removed \n from the elasticsearch query

Renamed file, for python script to work fine

=Only one script file was not refactored correctly

HadoopInputFormatIO with junits

Removed HadoopInputFormatIOConstants file and removed the use of constants from related classes.

Added Cassandra and Elastic tests/ITs on in HadoopInputFormatIO code.

Removed Writable Coder from HadoopInputFormat and referred it from Hadoop Common module

Restructured hadoop-input-format project for jdk1.8

Removed unwanted classes

Removed separate package of unit tests input, Moved them to inputformat package as per comments from iemejia

Removed .toDelete file and test.txt

Changes in resources files, added apache license , comments , renamed files as per Stephens comments, added comment to pom for child module

Created hifio child module for basic HadoopInputFormatIO code

Added dataflow runner profile in io pom

Updated teardown.sh

Removed the unused src folder

Updated version to 0.7.0

Removed the extra files created during rebase

Added comments in Kubernetes scripts

Cleaned up hifio pom.xml

Added while loop for sleep

Changes in poms- clean up, removal of version properties of sl4j, added df runner profile in hif test pom etc.

Reverted the change of Dataflow runner profile

Added seperate script for ycsb tool

Updated names of the script files

Added separate script for installing python

Removed <requireMavenVersion> tag from enforcer

Updated Javadoc

Removed the success message.

group id changed for parent as org.apache.beam

Refactoring -rename directory hadoop-input-format -> hadoop
 input-format,jdk1.8-tests

formatting correction for new module hadoop, renamed from hadoop-input-format to hadoop

Formatted the elasticsearch query

Added comment for pod initialization- increasing virtual memory

Updated teardown.sh scripts

Added comments in the scripts

Modified the comments

Changed the script to load deterministic data for Elasticsearch and changed Checksum

=Only one script file was not refactored correctly

Added show-health.sh to get elasticsearch cluster health info

Changes in comments

Change in Javadoc , change in number of fields in CassandraIT in transform function

ES scripts optimized, added in single file, separate clusters for IT and performance added

Removed tabs

emptyDir indentation rectified

Added comment for IPC_LOCK

Addressed Stephen's review comments 28thMar

Removed Cassandra LargeProductionCluster scripts from this PR

License added to 2 files, set -e added to few files

Made recordsReturned as AtomicLong to solve the Findbug issue of volatile long variable cannot be incremented

Changed to 1 replica for elasticsearch client Kubernetes script

Revert "Addressed Stephen's review comments 28thMar"

This reverts commit db207db9866fbd90f87cd02f7e3af46dea4d9912.

Revert "Modified the comments"

This reverts commit 8d389c6aae04bca3aad3f7c940935d6780441328.

Revert "ES scripts optimized, added in single file, separate clusters for IT and performance added"

This reverts commit 9c414cf395f8af95bc8cca9b08d33e701bc9bf82.

Commiting previous commit after reverting some unwanted jdbc files

Added property in pom for netty poll dependency

Changes in pom files as per Stephen's comments.
