+++ pom.xml
@@ -448,6 +448,12 @@
 
       <dependency>
         <groupId>org.apache.beam</groupId>
+        <artifactId>beam-sdks-java-io-hadoop-input-format</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+	
+      <dependency>
+        <groupId>org.apache.beam</groupId>
         <artifactId>beam-runners-core-construction-java</artifactId>
         <version>${project.version}</version>
       </dependency>
